Tracking now based on previous coordinates.
The search space for the template matching is now based on the last coordinates, so searching is much faster.  Parallel processing is removed now.

Potential issues are if the first coordinates are bad, or if the inventory is up for an extended period of time in a vehicle (this could cause us to go outside the search space).